feat: SJRA-147 classification criteria inheritances and pubmed ids as array
