+++ backend/docs/docs.go
@@ -6,7 +6,7 @@ import "github.com/swaggo/swag/v2"
 
 const docTemplate = `{
     "schemes": {{ marshal .Schemes }},
-    "components": {"schemas":{"types.Aggregation":{"description":"Aggregation represents an aggregation result","properties":{"count":{"description":"Count in the bucket","type":"integer"},"key":{"description":"Bucket key","type":"string"},"label":{"description":"Label corresponding to the key","type":"string"}},"type":"object"},"types.AggregationBodyWithSqon":{"properties":{"field":{"type":"string"},"size":{"type":"integer"},"sqon":{"$ref":"#/components/schemas/types.Sqon"}},"type":"object"},"types.ApiError":{"description":"ApiError represents an error","properties":{"detail":{},"message":{"type":"string"},"status":{"type":"integer"}},"required":["message","status"],"type":"object"},"types.AutoCompleteTerm":{"properties":{"highlight":{"$ref":"#/components/schemas/types.Term"},"source":{"$ref":"#/components/schemas/types.Term"}},"type":"object"},"types.AutocompleteResult":{"properties":{"type":{"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},"types.CaseFilters":{"properties":{"case_analysis":{"items":{"$ref":"#/components/schemas/types.Aggregation"},"type":"array","uniqueItems":false},"performer_lab":{"items":{"$ref":"#/components/schemas/types.Aggregation"},"type":"array","uniqueItems":false},"priority":{"items":{"$ref":"#/components/schemas/types.Aggregation"},"type":"array","uniqueItems":false},"project":{"items":{"$ref":"#/components/schemas/types.Aggregation"},"type":"array","uniqueItems":false},"requested_by":{"items":{"$ref":"#/components/schemas/types.Aggregation"},"type":"array","uniqueItems":false},"status":{"items":{"$ref":"#/components/schemas/types.Aggregation"},"type":"array","uniqueItems":false}},"required":["case_analysis","performer_lab","priority","project","requested_by","status"],"type":"object"},"types.CaseResult":{"description":"Line represented a case in case list","properties":{"case_analysis_code":{"type":"string"},"case_analysis_name":{"type":"string"},"case_analysis_type_code":{"type":"string"},"case_id":{"type":"integer"},"created_on":{"type":"string"},"managing_organization_code":{"type":"string"},"managing_organization_name":{"type":"string"},"mrn":{"type":"string"},"patient_id":{"type":"integer"},"performer_lab_code":{"type":"string"},"performer_lab_name":{"type":"string"},"prescriber":{"type":"string"},"primary_condition":{"type":"string"},"priority_code":{"type":"string"},"project_code":{"type":"string"},"request_id":{"type":"integer"},"requested_by_code":{"type":"string"},"requested_by_name":{"type":"string"},"status_code":{"type":"string"},"updated_on":{"type":"string"}},"required":["case_analysis_code","case_analysis_name","case_analysis_type_code","case_id","created_on","mrn","priority_code","project_code","requested_by_code","requested_by_name","status_code","updated_on"],"type":"object"},"types.Count":{"description":"Count represents count result","properties":{"count":{"description":"Number of results","type":"integer"}},"type":"object"},"types.CountBodyWithCriteria":{"properties":{"search_criteria":{"items":{"$ref":"#/components/schemas/types.SearchCriterion"},"type":"array","uniqueItems":false}},"type":"object"},"types.CountBodyWithSqon":{"properties":{"sqon":{"$ref":"#/components/schemas/types.Sqon"}},"type":"object"},"types.ExpendedOccurrence":{"properties":{"aa_change":{"type":"string"},"ad_alt":{"type":"integer"},"ad_total":{"type":"integer"},"af":{"description":"TODO","type":"number"},"cadd_phred":{"type":"number"},"cadd_score":{"type":"number"},"chromosome":{"type":"string"},"clinvar":{"items":{"type":"string"},"type":"array","uniqueItems":false},"dann_score":{"type":"number"},"dna_change":{"type":"string"},"end":{"type":"integer"},"exon_rank":{"type":"integer"},"exon_total":{"type":"integer"},"father_calls":{"items":{"type":"integer"},"type":"array","uniqueItems":false},"fathmm_pred":{"type":"string"},"fathmm_score":{"type":"number"},"filter":{"type":"string"},"genotype_quality":{"type":"integer"},"gnomad_loeuf":{"type":"number"},"gnomad_pli":{"type":"number"},"gnomad_v3_af":{"type":"number"},"hgvsg":{"type":"string"},"is_canonical":{"type":"boolean"},"is_mane_plus":{"type":"boolean"},"is_mane_select":{"type":"boolean"},"locus_id":{"type":"string"},"mother_calls":{"items":{"type":"integer"},"type":"array","uniqueItems":false},"omim_conditions":{"items":{"$ref":"#/components/schemas/types.OmimGenePanel"},"type":"array","uniqueItems":false},"parental_origin":{"type":"string"},"pf_wgs":{"type":"number"},"picked_consequences":{"items":{"type":"string"},"type":"array","uniqueItems":false},"qd":{"type":"number"},"revel_score":{"type":"number"},"rsnumber":{"type":"string"},"sift_pred":{"type":"string"},"sift_score":{"type":"number"},"spliceai_ds":{"type":"number"},"spliceai_type":{"items":{"type":"string"},"type":"array","uniqueItems":false},"start":{"type":"integer"},"symbol":{"type":"string"},"transcript_id":{"type":"string"},"transmission":{"type":"string"},"vep_impact":{"$ref":"#/components/schemas/types.VepImpact"},"zygosity":{"type":"string"}},"required":["genotype_quality","gnomad_v3_af","hgvsg","locus_id","picked_consequences"],"type":"object"},"types.FiltersBodyWithCriteria":{"properties":{"search_criteria":{"items":{"$ref":"#/components/schemas/types.SearchCriterion"},"type":"array","uniqueItems":false}},"type":"object"},"types.InterpretationGermline":{"properties":{"classification":{"type":"string"},"classification_criterias":{"items":{"type":"string"},"type":"array","uniqueItems":false},"condition":{"type":"string"},"created_at":{"type":"string"},"created_by":{"type":"string"},"created_by_name":{"type":"string"},"id":{"type":"string"},"interpretation":{"type":"string"},"locus_id":{"type":"string"},"metadata":{"$ref":"#/components/schemas/types.InterpretationMetadata"},"pubmed":{"items":{"$ref":"#/components/schemas/types.InterpretationPubmed"},"type":"array","uniqueItems":false},"sequencing_id":{"type":"string"},"transcript_id":{"type":"string"},"transmission_modes":{"items":{"type":"string"},"type":"array","uniqueItems":false},"updated_at":{"type":"string"},"updated_by":{"type":"string"},"updated_by_name":{"type":"string"}},"type":"object"},"types.InterpretationMetadata":{"properties":{"analysis_id":{"type":"string"},"patient_id":{"type":"string"},"variant_hash":{"type":"string"}},"type":"object"},"types.InterpretationPubmed":{"properties":{"citation":{"type":"string"},"citation_id":{"type":"string"}},"type":"object"},"types.InterpretationSomatic":{"properties":{"clinical_utility":{"type":"string"},"created_at":{"type":"string"},"created_by":{"type":"string"},"created_by_name":{"type":"string"},"id":{"type":"string"},"interpretation":{"type":"string"},"locus_id":{"type":"string"},"metadata":{"$ref":"#/components/schemas/types.InterpretationMetadata"},"oncogenicity":{"type":"string"},"oncogenicity_classification_criterias":{"items":{"type":"string"},"type":"array","uniqueItems":false},"pubmed":{"items":{"$ref":"#/components/schemas/types.InterpretationPubmed"},"type":"array","uniqueItems":false},"sequencing_id":{"type":"string"},"transcript_id":{"type":"string"},"tumoral_type":{"type":"string"},"updated_at":{"type":"string"},"updated_by":{"type":"string"},"updated_by_name":{"type":"string"}},"type":"object"},"types.LeafContent":{"properties":{"field":{"type":"string"},"value":{"items":{},"type":"array","uniqueItems":false}},"type":"object"},"types.ListBodyWithCriteria":{"description":"Body of a list request with search criteria","properties":{"additional_fields":{"items":{"type":"string"},"type":"array","uniqueItems":false},"limit":{"type":"integer"},"offset":{"type":"integer"},"page_index":{"type":"integer"},"search_criteria":{"items":{"$ref":"#/components/schemas/types.SearchCriterion"},"type":"array","uniqueItems":false},"sort":{"items":{"$ref":"#/components/schemas/types.SortBody"},"type":"array","uniqueItems":false}},"type":"object"},"types.ListBodyWithSqon":{"description":"Body of a list request","properties":{"additional_fields":{"items":{"type":"string"},"type":"array","uniqueItems":false},"limit":{"type":"integer"},"offset":{"type":"integer"},"page_index":{"type":"integer"},"sort":{"items":{"$ref":"#/components/schemas/types.SortBody"},"type":"array","uniqueItems":false},"sqon":{"$ref":"#/components/schemas/types.Sqon"}},"type":"object"},"types.Occurrence":{"description":"Occurrence represents an occurrence","properties":{"aa_change":{"type":"string"},"ad_ratio":{"type":"number"},"chromosome":{"type":"string"},"clinvar":{"items":{"type":"string"},"type":"array","uniqueItems":false},"filter":{"type":"string"},"genotype_quality":{"type":"integer"},"gnomad_v3_af":{"type":"number"},"hgvsg":{"type":"string"},"impact_score":{"type":"string"},"is_canonical":{"type":"boolean"},"is_mane_select":{"type":"boolean"},"locus_id":{"type":"string"},"omim_inheritance_code":{"items":{"type":"string"},"type":"array","uniqueItems":false},"pc_wgs":{"type":"integer"},"pf_wgs":{"type":"number"},"picked_consequences":{"items":{"type":"string"},"type":"array","uniqueItems":false},"pn_wgs":{"type":"integer"},"rsnumber":{"type":"string"},"seq_id":{"type":"integer"},"start":{"type":"integer"},"symbol":{"type":"string"},"task_id":{"type":"integer"},"transcript_id":{"type":"string"},"variant_class":{"type":"string"},"vep_impact":{"$ref":"#/components/schemas/types.VepImpact"},"zygosity":{"type":"string"}},"required":["ad_ratio","chromosome","genotype_quality","gnomad_v3_af","hgvsg","locus_id","pf_wgs","picked_consequences","seq_id","start","task_id","variant_class","zygosity"],"type":"object"},"types.OmimGenePanel":{"properties":{"inheritance_code":{"items":{"type":"string"},"type":"array","uniqueItems":false},"omim_phenotype_id":{"type":"string"},"panel":{"type":"string"}},"type":"object"},"types.PubmedCitation":{"properties":{"id":{"type":"string"},"nlm":{"$ref":"#/components/schemas/types.PubmedCitationDetails"}},"type":"object"},"types.PubmedCitationDetails":{"properties":{"format":{"type":"string"}},"type":"object"},"types.SearchCriterion":{"properties":{"field":{"type":"string"},"operator":{"type":"string"},"value":{"items":{},"type":"array","uniqueItems":false}},"type":"object"},"types.Sequencing":{"description":"Sequencing represents a sequencing","properties":{"affected_status":{"type":"string"},"analysis_type":{"type":"string"},"case_id":{"type":"integer"},"created_at":{"type":"string"},"experimental_strategy":{"type":"string"},"family_role":{"type":"string"},"part":{"type":"integer"},"patient_id":{"type":"integer"},"request_id":{"type":"integer"},"request_priority":{"type":"string"},"sample_id":{"type":"integer"},"seq_id":{"type":"integer"},"sex":{"type":"string"},"task_id":{"type":"integer"},"updated_at":{"type":"string"},"vcf_filepath":{"type":"string"}},"type":"object"},"types.SortBody":{"properties":{"field":{"type":"string"},"order":{"enum":["asc","desc"],"type":"string"}},"type":"object"},"types.Sqon":{"properties":{"content":{"oneOf":[{"$ref":"#/components/schemas/types.LeafContent"},{"items":{"$ref":"#/components/schemas/types.Sqon"},"type":"array"}]},"op":{"enum":["in","and","or","not","between","\u003e","\u003c","\u003e=","\u003c=","not-in","all"],"type":"string"}},"type":"object"},"types.Statistics":{"description":"Statistics represents statistics about a column","properties":{"max":{"description":"Max for numeric facet","type":"number"},"min":{"description":"Min for numeric facet","type":"number"}},"type":"object"},"types.StatisticsBodyWithSqon":{"properties":{"field":{"type":"string"},"sqon":{"$ref":"#/components/schemas/types.Sqon"}},"type":"object"},"types.Term":{"properties":{"id":{"type":"string"},"name":{"type":"string"},"term":{"type":"string"}},"type":"object"},"types.Transcript":{"properties":{"aa_change":{"type":"string"},"cadd_phred":{"type":"number"},"cadd_score":{"type":"number"},"consequences":{"items":{"type":"string"},"type":"array","uniqueItems":false},"dann_score":{"type":"number"},"dna_change":{"type":"string"},"exon_rank":{"type":"integer"},"exon_total":{"type":"integer"},"fathmm_pred":{"type":"string"},"fathmm_score":{"type":"number"},"is_canonical":{"type":"boolean"},"is_mane_plus":{"type":"boolean"},"is_mane_select":{"type":"boolean"},"lrt_pred":{"type":"string"},"lrt_score":{"type":"number"},"phyloP17way_primate":{"type":"number"},"polyphen2_hvar_pred":{"type":"string"},"polyphen2_hvar_score":{"type":"number"},"revel_score":{"type":"number"},"sift_pred":{"type":"string"},"sift_score":{"type":"number"},"transcript_id":{"type":"string"},"vep_impact":{"$ref":"#/components/schemas/types.VepImpact"}},"type":"object"},"types.UserSet":{"properties":{"active":{"type":"boolean"},"id":{"type":"string"},"ids":{"items":{"type":"string"},"type":"array","uniqueItems":false},"name":{"type":"string"},"type":{"type":"string"},"updated_at":{"type":"string"},"user_id":{"type":"string"}},"type":"object"},"types.VariantConsequence":{"properties":{"biotype":{"type":"string"},"gnomad_loeuf":{"type":"number"},"gnomad_pli":{"type":"number"},"is_picked":{"type":"boolean"},"spliceai_ds":{"type":"number"},"spliceai_type":{"items":{"type":"string"},"type":"array","uniqueItems":false},"symbol":{"type":"string"},"transcripts":{"items":{"$ref":"#/components/schemas/types.Transcript"},"type":"array","uniqueItems":false}},"type":"object"},"types.VariantExpendedInterpretedCase":{"properties":{"classification_criterias":{"type":"string"},"gene_symbol":{"type":"string"},"inheritances":{"type":"string"},"interpretation":{"type":"string"},"interpreter_name":{"type":"string"},"patient_id":{"type":"integer"},"patient_sex_code":{"type":"string"},"pubmed_ids":{"type":"string"}},"required":["classification_criterias","gene_symbol","inheritances","interpretation","interpreter_name","patient_id","patient_sex_code","pubmed_ids"],"type":"object"},"types.VariantHeader":{"properties":{"assembly_version":{"type":"string"},"hgvsg":{"type":"string"},"source":{"items":{"type":"string"},"type":"array","uniqueItems":false}},"required":["hgvsg"],"type":"object"},"types.VariantInterpretedCase":{"properties":{"case_analysis_code":{"type":"string"},"case_analysis_name":{"type":"string"},"case_id":{"type":"integer"},"classification":{"type":"string"},"condition_id":{"type":"string"},"condition_name":{"type":"string"},"interpretation_updated_on":{"type":"string"},"performer_lab_code":{"type":"string"},"performer_lab_name":{"type":"string"},"seq_id":{"type":"integer"},"status_code":{"type":"string"},"transcript_id":{"type":"string"},"zygosity":{"type":"string"}},"required":["case_analysis_code","case_analysis_name","case_id","classification","condition_id","condition_name","interpretation_updated_on","performer_lab_code","performer_lab_name","seq_id","status_code","transcript_id","zygosity"],"type":"object"},"types.VariantOverview":{"properties":{"aa_change":{"type":"string"},"cadd_phred":{"type":"number"},"cadd_score":{"type":"number"},"clinvar":{"items":{"type":"string"},"type":"array","uniqueItems":false},"clinvar_name":{"type":"string"},"dann_score":{"type":"number"},"dna_change":{"type":"string"},"exon_rank":{"type":"integer"},"exon_total":{"type":"integer"},"fathmm_pred":{"type":"string"},"fathmm_score":{"type":"number"},"gnomad_loeuf":{"type":"number"},"gnomad_pli":{"type":"number"},"gnomad_v3_af":{"type":"number"},"is_canonical":{"type":"boolean"},"is_mane_select":{"type":"boolean"},"locus":{"type":"string"},"lrt_pred":{"type":"string"},"lrt_score":{"type":"number"},"omim_conditions":{"items":{"$ref":"#/components/schemas/types.OmimGenePanel"},"type":"array","uniqueItems":false},"pc_wgs":{"type":"integer"},"pf_wgs":{"type":"number"},"phyloP17way_primate":{"type":"number"},"picked_consequences":{"items":{"type":"string"},"type":"array","uniqueItems":false},"pn_wgs":{"type":"integer"},"polyphen2_hvar_pred":{"type":"string"},"polyphen2_hvar_score":{"type":"number"},"revel_score":{"type":"number"},"rsnumber":{"type":"string"},"sift_pred":{"type":"string"},"sift_score":{"type":"number"},"spliceai_ds":{"type":"number"},"spliceai_type":{"items":{"type":"string"},"type":"array","uniqueItems":false},"symbol":{"type":"string"},"transcript_id":{"type":"string"},"vep_impact":{"$ref":"#/components/schemas/types.VepImpact"}},"required":["gnomad_v3_af","locus","pf_wgs","picked_consequences"],"type":"object"},"types.VariantUninterpretedCase":{"properties":{"case_analysis_code":{"type":"string"},"case_analysis_name":{"type":"string"},"case_id":{"type":"integer"},"created_on":{"type":"string"},"performer_lab_code":{"type":"string"},"performer_lab_name":{"type":"string"},"primary_condition_id":{"type":"string"},"primary_condition_name":{"type":"string"},"status_code":{"type":"string"},"updated_on":{"type":"string"},"zygosity":{"type":"string"}},"required":["case_analysis_code","case_analysis_name","case_id","created_on","performer_lab_code","performer_lab_name","primary_condition_id","primary_condition_name","status_code","updated_on","zygosity"],"type":"object"},"types.VepImpact":{"description":"TODO","enum":["MODIFIER","LOW","MODERATE","HIGH"],"type":"string","x-enum-varnames":["MODIFIER","LOW","MODERATE","HIGH"]}},"securitySchemes":{"bearerauth":{"bearerFormat":"JWT","scheme":"bearer","type":"http"}}},
+    "components": {"schemas":{"types.Aggregation":{"description":"Aggregation represents an aggregation result","properties":{"count":{"description":"Count in the bucket","type":"integer"},"key":{"description":"Bucket key","type":"string"},"label":{"description":"Label corresponding to the key","type":"string"}},"type":"object"},"types.AggregationBodyWithSqon":{"properties":{"field":{"type":"string"},"size":{"type":"integer"},"sqon":{"$ref":"#/components/schemas/types.Sqon"}},"type":"object"},"types.ApiError":{"description":"ApiError represents an error","properties":{"detail":{},"message":{"type":"string"},"status":{"type":"integer"}},"required":["message","status"],"type":"object"},"types.AutoCompleteTerm":{"properties":{"highlight":{"$ref":"#/components/schemas/types.Term"},"source":{"$ref":"#/components/schemas/types.Term"}},"type":"object"},"types.AutocompleteResult":{"properties":{"type":{"type":"string"},"value":{"type":"string"}},"required":["type","value"],"type":"object"},"types.CaseFilters":{"properties":{"case_analysis":{"items":{"$ref":"#/components/schemas/types.Aggregation"},"type":"array","uniqueItems":false},"performer_lab":{"items":{"$ref":"#/components/schemas/types.Aggregation"},"type":"array","uniqueItems":false},"priority":{"items":{"$ref":"#/components/schemas/types.Aggregation"},"type":"array","uniqueItems":false},"project":{"items":{"$ref":"#/components/schemas/types.Aggregation"},"type":"array","uniqueItems":false},"requested_by":{"items":{"$ref":"#/components/schemas/types.Aggregation"},"type":"array","uniqueItems":false},"status":{"items":{"$ref":"#/components/schemas/types.Aggregation"},"type":"array","uniqueItems":false}},"required":["case_analysis","performer_lab","priority","project","requested_by","status"],"type":"object"},"types.CaseResult":{"description":"Line represented a case in case list","properties":{"case_analysis_code":{"type":"string"},"case_analysis_name":{"type":"string"},"case_analysis_type_code":{"type":"string"},"case_id":{"type":"integer"},"created_on":{"type":"string"},"managing_organization_code":{"type":"string"},"managing_organization_name":{"type":"string"},"mrn":{"type":"string"},"patient_id":{"type":"integer"},"performer_lab_code":{"type":"string"},"performer_lab_name":{"type":"string"},"prescriber":{"type":"string"},"primary_condition":{"type":"string"},"priority_code":{"type":"string"},"project_code":{"type":"string"},"request_id":{"type":"integer"},"requested_by_code":{"type":"string"},"requested_by_name":{"type":"string"},"status_code":{"type":"string"},"updated_on":{"type":"string"}},"required":["case_analysis_code","case_analysis_name","case_analysis_type_code","case_id","created_on","mrn","priority_code","project_code","requested_by_code","requested_by_name","status_code","updated_on"],"type":"object"},"types.Count":{"description":"Count represents count result","properties":{"count":{"description":"Number of results","type":"integer"}},"type":"object"},"types.CountBodyWithCriteria":{"properties":{"search_criteria":{"items":{"$ref":"#/components/schemas/types.SearchCriterion"},"type":"array","uniqueItems":false}},"type":"object"},"types.CountBodyWithSqon":{"properties":{"sqon":{"$ref":"#/components/schemas/types.Sqon"}},"type":"object"},"types.ExpendedOccurrence":{"properties":{"aa_change":{"type":"string"},"ad_alt":{"type":"integer"},"ad_total":{"type":"integer"},"af":{"description":"TODO","type":"number"},"cadd_phred":{"type":"number"},"cadd_score":{"type":"number"},"chromosome":{"type":"string"},"clinvar":{"items":{"type":"string"},"type":"array","uniqueItems":false},"dann_score":{"type":"number"},"dna_change":{"type":"string"},"end":{"type":"integer"},"exon_rank":{"type":"integer"},"exon_total":{"type":"integer"},"father_calls":{"items":{"type":"integer"},"type":"array","uniqueItems":false},"fathmm_pred":{"type":"string"},"fathmm_score":{"type":"number"},"filter":{"type":"string"},"genotype_quality":{"type":"integer"},"gnomad_loeuf":{"type":"number"},"gnomad_pli":{"type":"number"},"gnomad_v3_af":{"type":"number"},"hgvsg":{"type":"string"},"is_canonical":{"type":"boolean"},"is_mane_plus":{"type":"boolean"},"is_mane_select":{"type":"boolean"},"locus_id":{"type":"string"},"mother_calls":{"items":{"type":"integer"},"type":"array","uniqueItems":false},"omim_conditions":{"items":{"$ref":"#/components/schemas/types.OmimGenePanel"},"type":"array","uniqueItems":false},"parental_origin":{"type":"string"},"pf_wgs":{"type":"number"},"picked_consequences":{"items":{"type":"string"},"type":"array","uniqueItems":false},"qd":{"type":"number"},"revel_score":{"type":"number"},"rsnumber":{"type":"string"},"sift_pred":{"type":"string"},"sift_score":{"type":"number"},"spliceai_ds":{"type":"number"},"spliceai_type":{"items":{"type":"string"},"type":"array","uniqueItems":false},"start":{"type":"integer"},"symbol":{"type":"string"},"transcript_id":{"type":"string"},"transmission":{"type":"string"},"vep_impact":{"$ref":"#/components/schemas/types.VepImpact"},"zygosity":{"type":"string"}},"required":["genotype_quality","gnomad_v3_af","hgvsg","locus_id","picked_consequences"],"type":"object"},"types.FiltersBodyWithCriteria":{"properties":{"search_criteria":{"items":{"$ref":"#/components/schemas/types.SearchCriterion"},"type":"array","uniqueItems":false}},"type":"object"},"types.InterpretationGermline":{"properties":{"classification":{"type":"string"},"classification_criterias":{"items":{"type":"string"},"type":"array","uniqueItems":false},"condition":{"type":"string"},"created_at":{"type":"string"},"created_by":{"type":"string"},"created_by_name":{"type":"string"},"id":{"type":"string"},"interpretation":{"type":"string"},"locus_id":{"type":"string"},"metadata":{"$ref":"#/components/schemas/types.InterpretationMetadata"},"pubmed":{"items":{"$ref":"#/components/schemas/types.InterpretationPubmed"},"type":"array","uniqueItems":false},"sequencing_id":{"type":"string"},"transcript_id":{"type":"string"},"transmission_modes":{"items":{"type":"string"},"type":"array","uniqueItems":false},"updated_at":{"type":"string"},"updated_by":{"type":"string"},"updated_by_name":{"type":"string"}},"type":"object"},"types.InterpretationMetadata":{"properties":{"analysis_id":{"type":"string"},"patient_id":{"type":"string"},"variant_hash":{"type":"string"}},"type":"object"},"types.InterpretationPubmed":{"properties":{"citation":{"type":"string"},"citation_id":{"type":"string"}},"type":"object"},"types.InterpretationSomatic":{"properties":{"clinical_utility":{"type":"string"},"created_at":{"type":"string"},"created_by":{"type":"string"},"created_by_name":{"type":"string"},"id":{"type":"string"},"interpretation":{"type":"string"},"locus_id":{"type":"string"},"metadata":{"$ref":"#/components/schemas/types.InterpretationMetadata"},"oncogenicity":{"type":"string"},"oncogenicity_classification_criterias":{"items":{"type":"string"},"type":"array","uniqueItems":false},"pubmed":{"items":{"$ref":"#/components/schemas/types.InterpretationPubmed"},"type":"array","uniqueItems":false},"sequencing_id":{"type":"string"},"transcript_id":{"type":"string"},"tumoral_type":{"type":"string"},"updated_at":{"type":"string"},"updated_by":{"type":"string"},"updated_by_name":{"type":"string"}},"type":"object"},"types.LeafContent":{"properties":{"field":{"type":"string"},"value":{"items":{},"type":"array","uniqueItems":false}},"type":"object"},"types.ListBodyWithCriteria":{"description":"Body of a list request with search criteria","properties":{"additional_fields":{"items":{"type":"string"},"type":"array","uniqueItems":false},"limit":{"type":"integer"},"offset":{"type":"integer"},"page_index":{"type":"integer"},"search_criteria":{"items":{"$ref":"#/components/schemas/types.SearchCriterion"},"type":"array","uniqueItems":false},"sort":{"items":{"$ref":"#/components/schemas/types.SortBody"},"type":"array","uniqueItems":false}},"type":"object"},"types.ListBodyWithSqon":{"description":"Body of a list request","properties":{"additional_fields":{"items":{"type":"string"},"type":"array","uniqueItems":false},"limit":{"type":"integer"},"offset":{"type":"integer"},"page_index":{"type":"integer"},"sort":{"items":{"$ref":"#/components/schemas/types.SortBody"},"type":"array","uniqueItems":false},"sqon":{"$ref":"#/components/schemas/types.Sqon"}},"type":"object"},"types.Occurrence":{"description":"Occurrence represents an occurrence","properties":{"aa_change":{"type":"string"},"ad_ratio":{"type":"number"},"chromosome":{"type":"string"},"clinvar":{"items":{"type":"string"},"type":"array","uniqueItems":false},"filter":{"type":"string"},"genotype_quality":{"type":"integer"},"gnomad_v3_af":{"type":"number"},"hgvsg":{"type":"string"},"impact_score":{"type":"string"},"is_canonical":{"type":"boolean"},"is_mane_select":{"type":"boolean"},"locus_id":{"type":"string"},"omim_inheritance_code":{"items":{"type":"string"},"type":"array","uniqueItems":false},"pc_wgs":{"type":"integer"},"pf_wgs":{"type":"number"},"picked_consequences":{"items":{"type":"string"},"type":"array","uniqueItems":false},"pn_wgs":{"type":"integer"},"rsnumber":{"type":"string"},"seq_id":{"type":"integer"},"start":{"type":"integer"},"symbol":{"type":"string"},"task_id":{"type":"integer"},"transcript_id":{"type":"string"},"variant_class":{"type":"string"},"vep_impact":{"$ref":"#/components/schemas/types.VepImpact"},"zygosity":{"type":"string"}},"required":["ad_ratio","chromosome","genotype_quality","gnomad_v3_af","hgvsg","locus_id","pf_wgs","picked_consequences","seq_id","start","task_id","variant_class","zygosity"],"type":"object"},"types.OmimGenePanel":{"properties":{"inheritance_code":{"items":{"type":"string"},"type":"array","uniqueItems":false},"omim_phenotype_id":{"type":"string"},"panel":{"type":"string"}},"type":"object"},"types.PubmedCitation":{"properties":{"id":{"type":"string"},"nlm":{"$ref":"#/components/schemas/types.PubmedCitationDetails"}},"type":"object"},"types.PubmedCitationDetails":{"properties":{"format":{"type":"string"}},"type":"object"},"types.SearchCriterion":{"properties":{"field":{"type":"string"},"operator":{"type":"string"},"value":{"items":{},"type":"array","uniqueItems":false}},"type":"object"},"types.Sequencing":{"description":"Sequencing represents a sequencing","properties":{"affected_status":{"type":"string"},"analysis_type":{"type":"string"},"case_id":{"type":"integer"},"created_at":{"type":"string"},"experimental_strategy":{"type":"string"},"family_role":{"type":"string"},"part":{"type":"integer"},"patient_id":{"type":"integer"},"request_id":{"type":"integer"},"request_priority":{"type":"string"},"sample_id":{"type":"integer"},"seq_id":{"type":"integer"},"sex":{"type":"string"},"task_id":{"type":"integer"},"updated_at":{"type":"string"},"vcf_filepath":{"type":"string"}},"type":"object"},"types.SortBody":{"properties":{"field":{"type":"string"},"order":{"enum":["asc","desc"],"type":"string"}},"type":"object"},"types.Sqon":{"properties":{"content":{"oneOf":[{"$ref":"#/components/schemas/types.LeafContent"},{"items":{"$ref":"#/components/schemas/types.Sqon"},"type":"array"}]},"op":{"enum":["in","and","or","not","between","\u003e","\u003c","\u003e=","\u003c=","not-in","all"],"type":"string"}},"type":"object"},"types.Statistics":{"description":"Statistics represents statistics about a column","properties":{"max":{"description":"Max for numeric facet","type":"number"},"min":{"description":"Min for numeric facet","type":"number"}},"type":"object"},"types.StatisticsBodyWithSqon":{"properties":{"field":{"type":"string"},"sqon":{"$ref":"#/components/schemas/types.Sqon"}},"type":"object"},"types.Term":{"properties":{"id":{"type":"string"},"name":{"type":"string"},"term":{"type":"string"}},"type":"object"},"types.Transcript":{"properties":{"aa_change":{"type":"string"},"cadd_phred":{"type":"number"},"cadd_score":{"type":"number"},"consequences":{"items":{"type":"string"},"type":"array","uniqueItems":false},"dann_score":{"type":"number"},"dna_change":{"type":"string"},"exon_rank":{"type":"integer"},"exon_total":{"type":"integer"},"fathmm_pred":{"type":"string"},"fathmm_score":{"type":"number"},"is_canonical":{"type":"boolean"},"is_mane_plus":{"type":"boolean"},"is_mane_select":{"type":"boolean"},"lrt_pred":{"type":"string"},"lrt_score":{"type":"number"},"phyloP17way_primate":{"type":"number"},"polyphen2_hvar_pred":{"type":"string"},"polyphen2_hvar_score":{"type":"number"},"revel_score":{"type":"number"},"sift_pred":{"type":"string"},"sift_score":{"type":"number"},"transcript_id":{"type":"string"},"vep_impact":{"$ref":"#/components/schemas/types.VepImpact"}},"type":"object"},"types.UserSet":{"properties":{"active":{"type":"boolean"},"id":{"type":"string"},"ids":{"items":{"type":"string"},"type":"array","uniqueItems":false},"name":{"type":"string"},"type":{"type":"string"},"updated_at":{"type":"string"},"user_id":{"type":"string"}},"type":"object"},"types.VariantConsequence":{"properties":{"biotype":{"type":"string"},"gnomad_loeuf":{"type":"number"},"gnomad_pli":{"type":"number"},"is_picked":{"type":"boolean"},"spliceai_ds":{"type":"number"},"spliceai_type":{"items":{"type":"string"},"type":"array","uniqueItems":false},"symbol":{"type":"string"},"transcripts":{"items":{"$ref":"#/components/schemas/types.Transcript"},"type":"array","uniqueItems":false}},"type":"object"},"types.VariantExpendedInterpretedCase":{"properties":{"classification_criterias":{"items":{"type":"string"},"type":"array","uniqueItems":false},"gene_symbol":{"type":"string"},"inheritances":{"items":{"type":"string"},"type":"array","uniqueItems":false},"interpretation":{"type":"string"},"interpreter_name":{"type":"string"},"patient_id":{"type":"integer"},"patient_sex_code":{"type":"string"},"pubmed_ids":{"items":{"type":"string"},"type":"array","uniqueItems":false}},"required":["classification_criterias","gene_symbol","inheritances","interpretation","interpreter_name","patient_id","patient_sex_code","pubmed_ids"],"type":"object"},"types.VariantHeader":{"properties":{"assembly_version":{"type":"string"},"hgvsg":{"type":"string"},"source":{"items":{"type":"string"},"type":"array","uniqueItems":false}},"required":["hgvsg"],"type":"object"},"types.VariantInterpretedCase":{"properties":{"case_analysis_code":{"type":"string"},"case_analysis_name":{"type":"string"},"case_id":{"type":"integer"},"classification":{"type":"string"},"condition_id":{"type":"string"},"condition_name":{"type":"string"},"interpretation_updated_on":{"type":"string"},"performer_lab_code":{"type":"string"},"performer_lab_name":{"type":"string"},"seq_id":{"type":"integer"},"status_code":{"type":"string"},"transcript_id":{"type":"string"},"zygosity":{"type":"string"}},"required":["case_analysis_code","case_analysis_name","case_id","classification","condition_id","condition_name","interpretation_updated_on","performer_lab_code","performer_lab_name","seq_id","status_code","transcript_id","zygosity"],"type":"object"},"types.VariantOverview":{"properties":{"aa_change":{"type":"string"},"cadd_phred":{"type":"number"},"cadd_score":{"type":"number"},"clinvar":{"items":{"type":"string"},"type":"array","uniqueItems":false},"clinvar_name":{"type":"string"},"dann_score":{"type":"number"},"dna_change":{"type":"string"},"exon_rank":{"type":"integer"},"exon_total":{"type":"integer"},"fathmm_pred":{"type":"string"},"fathmm_score":{"type":"number"},"gnomad_loeuf":{"type":"number"},"gnomad_pli":{"type":"number"},"gnomad_v3_af":{"type":"number"},"is_canonical":{"type":"boolean"},"is_mane_select":{"type":"boolean"},"locus":{"type":"string"},"lrt_pred":{"type":"string"},"lrt_score":{"type":"number"},"omim_conditions":{"items":{"$ref":"#/components/schemas/types.OmimGenePanel"},"type":"array","uniqueItems":false},"pc_wgs":{"type":"integer"},"pf_wgs":{"type":"number"},"phyloP17way_primate":{"type":"number"},"picked_consequences":{"items":{"type":"string"},"type":"array","uniqueItems":false},"pn_wgs":{"type":"integer"},"polyphen2_hvar_pred":{"type":"string"},"polyphen2_hvar_score":{"type":"number"},"revel_score":{"type":"number"},"rsnumber":{"type":"string"},"sift_pred":{"type":"string"},"sift_score":{"type":"number"},"spliceai_ds":{"type":"number"},"spliceai_type":{"items":{"type":"string"},"type":"array","uniqueItems":false},"symbol":{"type":"string"},"transcript_id":{"type":"string"},"vep_impact":{"$ref":"#/components/schemas/types.VepImpact"}},"required":["gnomad_v3_af","locus","pf_wgs","picked_consequences"],"type":"object"},"types.VariantUninterpretedCase":{"properties":{"case_analysis_code":{"type":"string"},"case_analysis_name":{"type":"string"},"case_id":{"type":"integer"},"created_on":{"type":"string"},"performer_lab_code":{"type":"string"},"performer_lab_name":{"type":"string"},"primary_condition_id":{"type":"string"},"primary_condition_name":{"type":"string"},"status_code":{"type":"string"},"updated_on":{"type":"string"},"zygosity":{"type":"string"}},"required":["case_analysis_code","case_analysis_name","case_id","created_on","performer_lab_code","performer_lab_name","primary_condition_id","primary_condition_name","status_code","updated_on","zygosity"],"type":"object"},"types.VepImpact":{"description":"TODO","enum":["MODIFIER","LOW","MODERATE","HIGH"],"type":"string","x-enum-varnames":["MODIFIER","LOW","MODERATE","HIGH"]}},"securitySchemes":{"bearerauth":{"bearerFormat":"JWT","scheme":"bearer","type":"http"}}},
     "info": {"description":"{{escape .Description}}","license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"},"title":"{{.Title}}","version":"{{.Version}}"},
     "externalDocs": {"description":"","url":""},
     "paths": {"/cases/autocomplete":{"get":{"description":"Retrieve types.AutocompleteResult list of ids matching prefix","operationId":"autocompleteCases","parameters":[{"description":"Prefix","in":"query","name":"prefix","required":true,"schema":{"type":"string"}},{"description":"Limit","in":"query","name":"limit","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/types.AutocompleteResult"},"type":"array"}}},"description":"OK"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get types.AutocompleteResult list of matching prefix","tags":["cases"]}},"/cases/count":{"post":{"description":"Counts cases","operationId":"countCases","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.CountBodyWithCriteria"}}},"description":"Count Body","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.Count"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Count cases","tags":["cases"]}},"/cases/filters":{"post":{"description":"Retrieve types.CaseFilters cases filters","operationId":"casesFilters","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.FiltersBodyWithCriteria"}}},"description":"Filters Body","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.CaseFilters"}}},"description":"OK"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get types.CaseFilters cases filters","tags":["cases"]}},"/cases/list":{"post":{"description":"List cases","operationId":"listCases","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ListBodyWithCriteria"}}},"description":"List Body","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/types.CaseResult"},"type":"array"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"List cases","tags":["cases"]}},"/hpo/autocomplete":{"get":{"description":"Retrieve AutoCompleteTerm list of HPO terms matching input string with highlighted","operationId":"hpoTermAutoComplete","parameters":[{"description":"Prefix","in":"query","name":"prefix","required":true,"schema":{"type":"string"}},{"description":"Limit","in":"query","name":"limit","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/types.AutoCompleteTerm"},"type":"array"}}},"description":"OK"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get AutoCompleteTerm list of matching input string with highlighted","tags":["hpo"]}},"/interpretations/germline":{"get":{"description":"Search interpretation germline","operationId":"SearchInterpretationGermline","responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/types.InterpretationGermline"},"type":"array"}}},"description":"OK"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Search interpretation germline","tags":["interpretations"]}},"/interpretations/germline/{sequencing_id}/{locus_id}/{transcript_id}":{"get":{"description":"Get interpretation germline","operationId":"GetInterpretationGermline","parameters":[{"description":"Sequencing ID","in":"path","name":"sequencing_id","required":true,"schema":{"type":"string"}},{"description":"Locus ID","in":"path","name":"locus_id","required":true,"schema":{"type":"string"}},{"description":"Transcript ID","in":"path","name":"transcript_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.InterpretationGermline"}}},"description":"OK"},"206":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.InterpretationGermline"}}},"description":"Partial Content"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get interpretation germline","tags":["interpretations"]},"post":{"description":"Create or Update interpretation germline","operationId":"PostInterpretationGermline","parameters":[{"description":"Sequencing ID","in":"path","name":"sequencing_id","required":true,"schema":{"type":"string"}},{"description":"Locus ID","in":"path","name":"locus_id","required":true,"schema":{"type":"string"}},{"description":"Transcript ID","in":"path","name":"transcript_id","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.InterpretationGermline"}}},"description":"Interpretation Body","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.InterpretationGermline"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Create or Update interpretation germline","tags":["interpretations"]}},"/interpretations/pubmed/{citation_id}":{"get":{"description":"Get pubmed citation by ID","operationId":"GetPubmedCitation","parameters":[{"description":"Citation ID","in":"path","name":"citation_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.PubmedCitation"}}},"description":"OK"},"206":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.PubmedCitation"}}},"description":"Partial Content"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get pubmed citation by ID","tags":["interpretations"]}},"/interpretations/somatic":{"get":{"description":"Search interpretation somatic","operationId":"SearchInterpretationSomatic","responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/types.InterpretationSomatic"},"type":"array"}}},"description":"OK"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Search interpretation somatic","tags":["interpretations"]}},"/interpretations/somatic/{sequencing_id}/{locus_id}/{transcript_id}":{"get":{"description":"Get interpretation somatic","operationId":"GetInterpretationSomatic","parameters":[{"description":"Sequencing ID","in":"path","name":"sequencing_id","required":true,"schema":{"type":"string"}},{"description":"Locus ID","in":"path","name":"locus_id","required":true,"schema":{"type":"string"}},{"description":"Transcript ID","in":"path","name":"transcript_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.InterpretationSomatic"}}},"description":"OK"},"206":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.InterpretationSomatic"}}},"description":"Partial Content"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get interpretation somatic","tags":["interpretations"]},"post":{"description":"Create or Update interpretation somatic","operationId":"PostInterpretationSomatic","parameters":[{"description":"Sequencing ID","in":"path","name":"sequencing_id","required":true,"schema":{"type":"string"}},{"description":"Locus ID","in":"path","name":"locus_id","required":true,"schema":{"type":"string"}},{"description":"Transcript ID","in":"path","name":"transcript_id","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.InterpretationSomatic"}}},"description":"Interpretation Body","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.InterpretationSomatic"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Create or Update interpretation somatic","tags":["interpretations"]}},"/mondo/autocomplete":{"get":{"description":"Retrieve AutoCompleteTerm list of mondo terms matching input string with highlighted","operationId":"mondoTermAutoComplete","parameters":[{"description":"Prefix","in":"query","name":"prefix","required":true,"schema":{"type":"string"}},{"description":"Limit","in":"query","name":"limit","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/types.AutoCompleteTerm"},"type":"array"}}},"description":"OK"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get AutoCompleteTerm list of matching input string with highlighted","tags":["mondo"]}},"/occurrences/germline/{seq_id}/aggregate":{"post":{"description":"Aggregate germline occurrences for a given sequence ID","operationId":"aggregateGermlineOccurrences","parameters":[{"description":"Sequence ID","in":"path","name":"seq_id","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.AggregationBodyWithSqon"}}},"description":"Aggregation Body","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/types.Aggregation"},"type":"array"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Aggregate germline occurrences","tags":["occurrences"]}},"/occurrences/germline/{seq_id}/count":{"post":{"description":"Counts germline occurrences for a given sequence ID","operationId":"countGermlineOccurrences","parameters":[{"description":"Sequence ID","in":"path","name":"seq_id","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.CountBodyWithSqon"}}},"description":"Count Body","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.Count"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Count germline occurrences","tags":["occurrences"]}},"/occurrences/germline/{seq_id}/list":{"post":{"description":"List germline occurrences for a given sequence ID","operationId":"listGermlineOccurrences","parameters":[{"description":"Sequence ID","in":"path","name":"seq_id","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ListBodyWithSqon"}}},"description":"List Body","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/types.Occurrence"},"type":"array"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"List germline occurrences","tags":["occurrences"]}},"/occurrences/germline/{seq_id}/statistics":{"post":{"description":"Return statistics about a field for a given sequence ID","operationId":"statisticsGermlineOccurrences","parameters":[{"description":"Sequence ID","in":"path","name":"seq_id","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.StatisticsBodyWithSqon"}}},"description":"Statistics Body","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.Statistics"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Statistics of germline occurrences","tags":["occurrences"]}},"/occurrences/germline/{seq_id}/{locus_id}/expended":{"get":{"description":"Retrieve ExpendedOccurrence data for a given locus ID","operationId":"getExpendedGermlineOccurrence","parameters":[{"description":"Sequence ID","in":"path","name":"seq_id","required":true,"schema":{"type":"string"}},{"description":"Locus ID","in":"path","name":"locus_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ExpendedOccurrence"}}},"description":"OK"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get a germline ExpendedOccurrence","tags":["occurrences"]}},"/sequencing/{seq_id}":{"get":{"description":"Retrieve Sequencing data for a given sequence ID","operationId":"getSequencing","parameters":[{"description":"Sequence ID","in":"path","name":"seq_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.Sequencing"}}},"description":"OK"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get a Sequencing","tags":["sequencing"]}},"/status":{"get":{"description":"Returns the current status of the API","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"OK"}},"summary":"Get API status","tags":["status"]}},"/users/sets/{user_set_id}":{"get":{"description":"Get user set","operationId":"GetUserSet","parameters":[{"description":"UserSet ID","in":"path","name":"user_set_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.UserSet"}}},"description":"OK"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get user set by id","tags":["user_sets"]}},"/variants/germline/{locus_id}/consequences":{"get":{"description":"Retrieve germline Variant Consequences for a given locus","operationId":"getGermlineVariantConsequences","parameters":[{"description":"Locus ID","in":"path","name":"locus_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/types.VariantConsequence"},"type":"array"}}},"description":"OK"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get list of VariantConsequences for a germline variant","tags":["variant"]}},"/variants/germline/{locus_id}/header":{"get":{"description":"Retrieve germline Variant Header data for a given locus","operationId":"getGermlineVariantHeader","parameters":[{"description":"Locus ID","in":"path","name":"locus_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.VariantHeader"}}},"description":"OK"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get a germline VariantHeader","tags":["variant"]}},"/variants/germline/{locus_id}/overview":{"get":{"description":"Retrieve germline Variant Overview data for a given locus","operationId":"getGermlineVariantOverview","parameters":[{"description":"Locus ID","in":"path","name":"locus_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.VariantOverview"}}},"description":"OK"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get a germline VariantOverview","tags":["variant"]}},"/variants/{locus_id}/cases/interpreted":{"post":{"description":"Retrieve Variant interpreted cases for a given locus","operationId":"getVariantInterpretedCases","parameters":[{"description":"Locus ID","in":"path","name":"locus_id","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ListBodyWithCriteria"}}},"description":"Filters Body","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/types.VariantInterpretedCase"},"type":"array"}}},"description":"OK"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get list of interpreted Cases for a variant","tags":["variant"]}},"/variants/{locus_id}/cases/interpreted/{seq_id}/{transcript_id}":{"get":{"description":"Retrieve expended interpreted case for a given locus, sequencing and transcript","operationId":"getExpendedVariantInterpretedCase","parameters":[{"description":"Locus ID","in":"path","name":"locus_id","required":true,"schema":{"type":"string"}},{"description":"Seq ID","in":"path","name":"seq_id","required":true,"schema":{"type":"string"}},{"description":"Transcript ID","in":"path","name":"transcript_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.VariantExpendedInterpretedCase"}}},"description":"OK"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get expended interpreted case for a given locus, sequencing and transcript","tags":["variant"]}},"/variants/{locus_id}/cases/uninterpreted":{"post":{"description":"Retrieve Variant uninterpreted cases for a given locus","operationId":"postVariantUninterpretedCases","parameters":[{"description":"Locus ID","in":"path","name":"locus_id","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.FiltersBodyWithCriteria"}}},"description":"Filters Body","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/types.VariantUninterpretedCase"},"type":"array"}}},"description":"OK"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/types.ApiError"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Get list of uninterpreted Cases for a variant","tags":["variant"]}}},
