+++ backend/internal/types/case.go
@@ -210,44 +210,3 @@ type CaseFilters struct {
 	PerformerLab []Aggregation `json:"performer_lab" validate:"required"`
 	RequestedBy  []Aggregation `json:"requested_by" validate:"required"`
 }
-
-type VariantInterpretedCase = struct {
-	SeqId                   int       `json:"seq_id" validate:"required"`
-	CaseId                  int       `json:"case_id" validate:"required"`
-	TranscriptId            string    `json:"transcript_id" validate:"required"`
-	InterpretationUpdatedOn time.Time `json:"interpretation_updated_on" validate:"required"`
-	ConditionId             string    `json:"condition_id" validate:"required"`
-	ConditionName           string    `json:"condition_name" validate:"required"`
-	Classification          string    `json:"classification" validate:"required"`
-	Zygosity                string    `json:"zygosity" validate:"required"`
-	PerformerLabCode        string    `json:"performer_lab_code" validate:"required"`
-	PerformerLabName        string    `json:"performer_lab_name" validate:"required"`
-	CaseAnalysisCode        string    `json:"case_analysis_code" validate:"required"`
-	CaseAnalysisName        string    `json:"case_analysis_name" validate:"required"`
-	StatusCode              string    `json:"status_code" validate:"required"`
-} // @name VariantInterpretedCase
-
-type VariantUninterpretedCase = struct {
-	CaseId               int       `json:"case_id" validate:"required"`
-	CreatedOn            time.Time `json:"created_on" validate:"required"`
-	UpdatedOn            time.Time `json:"updated_on" validate:"required"`
-	PrimaryConditionId   string    `json:"primary_condition_id" validate:"required"`
-	PrimaryConditionName string    `json:"primary_condition_name" validate:"required"`
-	Zygosity             string    `json:"zygosity" validate:"required"`
-	PerformerLabCode     string    `json:"performer_lab_code" validate:"required"`
-	PerformerLabName     string    `json:"performer_lab_name" validate:"required"`
-	CaseAnalysisCode     string    `json:"case_analysis_code" validate:"required"`
-	CaseAnalysisName     string    `json:"case_analysis_name" validate:"required"`
-	StatusCode           string    `json:"status_code" validate:"required"`
-} // @name VariantUninterpretedCase
-
-type VariantExpendedInterpretedCase = struct {
-	PatientID               int    `json:"patient_id" validate:"required"`
-	InterpreterName         string `json:"interpreter_name" validate:"required"`
-	Interpretation          string `json:"interpretation" validate:"required"`
-	GeneSymbol              string `json:"gene_symbol" validate:"required"`
-	ClassificationCriterias string `json:"classification_criterias" validate:"required"`
-	Inheritances            string `json:"inheritances" validate:"required"`
-	PatientSexCode          string `json:"patient_sex_code" validate:"required"`
-	PubmedIDs               string `json:"pubmed_ids" validate:"required"`
-} // @name VariantExpendedInterpretedCase
